Fixed priorities bug on import
RuneScape uses a priority footer flag of 255 for per-face model priorities. If this flag is some other number, say 10, the whole model's faces get a priority of 10. Before, it was importing no priorities for things like arms models which have a model-wide priority of 10. Now it works how it should on import and export!